Update np type

+++ hermes_eea/util/time/ccsds_to_cdf_time.py
@@ -44,7 +44,7 @@ def converting_ccsds_times_to_cdf(coarse, fine):
     p1 = np.int64(coarse) * np.int64(tai_time["nanosPerSec"])
     p2 = np.int64(fine) * np.int64(tai_time["nanosPerMicro"])
     epoch = p1 + p2
-    result = np.uint(epoch - tai_time["taiEpoch_tt2000"])
+    result = np.uint64(epoch - tai_time["taiEpoch_tt2000"])
     return result
